+++ lib/pagy.rb
@@ -13,17 +13,17 @@ class Pagy
   end
 
   # Default core vars: constant for easy access, but mutable for customizable defaults
-  DEFAULT = { page:         1, # rubocop:disable Style/MutableConstant
-              items:        20,
-              outset:       0,
-              size:         7,
-              page_param:   :page,
-              params:       {},
-              fragment:     '',
-              link_extra:   '',
-              i18n_key:     'pagy.item_name',
-              cycle:        false,
-              request_path: '' }
+  DEFAULT = { page:          1, # rubocop:disable Style/MutableConstant
+              items:         20,
+              outset:        0,
+              size:          7,
+              page_param:    :page,
+              params:        {},
+              fragment:      '',
+              link_extra:    '',
+              item_i18n_key: 'pagy.item_name',
+              cycle:         false,
+              request_path:  '' }
 
   attr_reader :count, :page, :items, :vars, :pages, :last, :offset, :in, :from, :to, :prev, :next, :params, :request_path