+++ lib/pagy/extras/items.rb
@@ -35,7 +35,7 @@ class Pagy # :nodoc:
       ITEMS_PLACEHOLDER = '__pagy_items__'
 
       # Return the items selector HTML. For example "Show [20] items per page"
-      def pagy_items_selector_js(pagy, pagy_id: nil, item_name: nil, i18n_key: nil, link_extra: '')
+      def pagy_items_selector_js(pagy, pagy_id: nil, item_name: nil, item_i18n_key: nil, link_extra: '')
         return '' unless pagy.vars[:items_extra]
 
         p_id           = %( id="#{pagy_id}") if pagy_id
@@ -48,9 +48,10 @@ class Pagy # :nodoc:
         html  = +%(<span#{p_id} class="pagy-items-selector-js" #{pagy_data(pagy, :selector, pagy.from, link)}>)
         input = %(<input type="number" min="1" max="#{p_vars[:max_items]}" value="#{
                     p_items}" style="padding: 0; text-align: center; width: #{p_items.to_s.length + 1}rem;">)
-        html << pagy_t('pagy.items_selector_js', item_name: item_name || pagy_t(i18n_key || p_vars[:i18n_key], count: p_items),
-                                                 items_input: input,
-                                                 count: p_items)
+        html << pagy_t('pagy.items_selector_js',
+                       item_name: item_name || pagy_t(item_i18n_key || p_vars[:item_i18n_key], count: p_items),
+                       items_input: input,
+                       count: p_items)
         html << %(</span>)
       end
     end